requirements.txt
Flask==3.0.2
requests==2.32.3
python-dotenv==1.0.1
flake8==7.1.0

